# internal/httpapi/admin/openapi/swagger.json
{
    "components": {
        "schemas": {
...
            "admin.changePasswordRequest": {
                "properties": {
                    "current_password": {
                        "type": "string"
                    },
                    "new_password": {
                        "minLength": 10,
                        "type": "string"
                    }
                },
                "required": [
                    "current_password",
                    "new_password"
                ],
                "type": "object"
            },
            "admin.createAPIKeyRequest": {
                "properties": {
                    "name": {
                        "maxLength": 100,
                        "type": "string"
                    }
                },
                "required": [
                    "name"
                ],
                "type": "object"
            },
...
            "admin.errorBody": {
                "properties": {
                    "details": {
                        "additionalProperties": {
                            "type": "string"
                        },
                        "type": "object"
                    },
                    "error": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "admin.loginRequest": {
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "totp_code": {
                        "type": "string"
                    }
                },
                "required": [
                    "email",
                    "password"
                ],
                "type": "object"
            },
            "admin.meResponse": {
                "properties": {
                    "csrf_token": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    },
                    "totp_enabled": {
                        "type": "boolean"
                    },
                    "via_api_key": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "admin.resetInput": {
                "properties": {
                    "password": {
                        "minLength": 10,
                        "type": "string"
                    },
                    "token": {
                        "type": "string"
                    }
                },
                "required": [
                    "password",
                    "token"
                ],
                "type": "object"
            },
...
            "admin.totpCodeRequest": {
                "properties": {
                    "code": {
                        "type": "string"
                    }
                },
                "required": [
                    "code"
                ],
                "type": "object"
            },
...
            "model.APIKey": {
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "key_prefix": {
                        "type": "string"
                    },
                    "last_used_at": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "revoked_at": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "description": "\"Bearer sbc_...\" API key, or the sbc_access cookie set by POST /auth/login (with X-CSRF-Token for state changes).",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "Session Border Controller control plane: customers, carriers, rate decks, routing, CDRs, reports.",
        "title": "SuperSBC Admin API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/auth/api-keys": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/model.APIKey"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "List API keys (own keys; admins see all)",
                "tags": [
                    "auth"
                ]
            },
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/admin.createAPIKeyRequest",
                                "summary": "body",
                                "description": "name"
                            }
                        }
                    },
                    "description": "name",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {},
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Created"
                    }
                },
                "summary": "Create an API key (the plaintext is returned once)",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/api-keys/{id}": {
            "delete": {
                "parameters": [
                    {
                        "description": "key id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "summary": "Revoke an API key",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/login": {
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/admin.loginRequest",
                                "summary": "body",
                                "description": "credentials"
                            }
                        }
                    },
                    "description": "credentials",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/admin.meResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/admin.errorBody"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "429": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/admin.errorBody"
                                }
                            }
                        },
                        "description": "Too Many Requests"
                    }
                },
                "summary": "Log in with email and password (and TOTP code when enabled)",
                "tags": [
                    "auth"
                ]
            }
        },
...
        "/auth/password": {
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/admin.changePasswordRequest",
                                "summary": "body",
                                "description": "passwords"
                            }
                        }
                    },
                    "description": "passwords",
                    "required": true
                },
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "summary": "Change own password (revokes other sessions)",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/refresh": {
            "post": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/admin.meResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/admin.errorBody"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "summary": "Rotate the refresh token and issue a new access token",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/reset": {
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/admin.resetInput",
                                "summary": "body",
                                "description": "token and password"
                            }
                        }
                    },
                    "description": "token and password",
                    "required": true
                },
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "summary": "Set a new password with a one-time reset token (unauthenticated)",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/totp/confirm": {
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/admin.totpCodeRequest",
                                "summary": "body",
                                "description": "code"
                            }
                        }
                    },
                    "description": "code",
                    "required": true
                },
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "summary": "Confirm a TOTP code and enable 2FA",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/totp/disable": {
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/admin.totpCodeRequest",
                                "summary": "body",
                                "description": "code"
                            }
                        }
                    },
                    "description": "code",
                    "required": true
                },
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "summary": "Disable 2FA (requires a valid code)",
                "tags": [
                    "auth"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/api/v1"
        }
    ]
}